+++ arbitrator/wasm-libraries/user-host/src/program.rs
@@ -1,15 +1,14 @@
 // Copyright 2022-2023, Offchain Labs, Inc.
 // For license information, see https://github.com/OffchainLabs/nitro/blob/master/LICENSE
 
-use std::fmt::Display;
-
-use crate::{evm_api::ApiCaller, host::UserHost};
+use crate::{evm_api::ApiCaller, traits::UserHost};
 use arbutil::{
     evm::{js::JsEvmApi, EvmData},
     wavm, Bytes20, Bytes32, Color,
 };
 use eyre::{eyre, Result};
 use prover::programs::prelude::*;
+use std::fmt::Display;
 
 /// The list of active programs. The current program is always the last.
 ///
